Rename and move sources
