+++ src/AzureDevopsAgentOperator/AgentOperatorBuilder.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace AzDoAgentDrainer
+namespace AzureDevopsAgentOperator
 {
     public class AgentOperatorBuilder
     {
